requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.22.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
pydantic>=2.0.0
python-multipart>=0.0.6
numpy>=1.24.0
aiohttp>=3.9.0
PyTurboJPEG<2.0
